# timelio-back/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /user/emplois/{idEmploi}/events/{idEvent}:
    get:
      tags:
      - evenements-controller
      operationId: getEvent
      parameters:
      - name: idEmploi
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: idEvent
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Evenement'
    put:
      tags:
      - evenements-controller
      operationId: majEvent
      parameters:
      - name: idEmploi
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: idEvent
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventForm'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Evenement'
    delete:
      tags:
      - evenements-controller
      operationId: supprimerEvent
      parameters:
      - name: idEmploi
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: idEvent
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
...
  /emplois/{codeAcces}/events/{idEvent}:
    get:
      tags:
      - evenements-controller
      operationId: getEvent_1
      parameters:
      - name: codeAcces
        in: path
        required: true
        schema:
          type: string
      - name: idEvent
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Evenement'
    put:
      tags:
      - evenements-controller
      operationId: majEvent_1
      parameters:
      - name: codeAcces
        in: path
        required: true
        schema:
          type: string
      - name: idEvent
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventForm'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Evenement'
    delete:
      tags:
      - evenements-controller
      operationId: supprimerEvent_1
      parameters:
      - name: codeAcces
        in: path
        required: true
        schema:
          type: string
      - name: idEvent
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
...
  /user/emplois/{id}/events:
    get:
      tags:
      - evenements-controller
      operationId: listerEvents
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: pagination
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageEvenement'
    post:
      tags:
      - evenements-controller
      operationId: creerEvent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventForm'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Evenement'
  /user/emplois/{idEmploi}/events/search:
    post:
      tags:
      - evenements-controller
      operationId: searchEvents
      parameters:
      - name: idEmploi
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriteresRerchercheEvents'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Evenement'
...
  /emplois/{codeAcces}/events:
    get:
      tags:
      - evenements-controller
      operationId: listerEvents_1
      parameters:
      - name: codeAcces
        in: path
        required: true
        schema:
          type: string
      - name: pagination
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageEvenement'
    post:
      tags:
      - evenements-controller
      operationId: creerEvent_1
      parameters:
      - name: codeAcces
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventForm'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Evenement'
  /emplois/{codeAcces}/events/search:
    post:
      tags:
      - evenements-controller
      operationId: searchEvents_1
      parameters:
      - name: codeAcces
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriteresRerchercheEvents'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Evenement'
...
components:
  schemas:
...
    EmploiTemps:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nom:
          type: string
        codeAcces:
          type: string
        publique:
          type: boolean
        owner:
          $ref: '#/components/schemas/Utilisateur'
        evenements:
          type: array
          items:
            $ref: '#/components/schemas/Evenement'
    Evenement:
      type: object
      properties:
        id:
          type: integer
          format: int64
        dateDebut:
          type: string
          format: date-time
        duree:
          type: object
          properties:
            seconds:
              type: integer
              format: int64
            nano:
              type: integer
              format: int32
            negative:
              type: boolean
            units:
              type: array
              items:
                type: object
                properties:
                  durationEstimated:
                    type: boolean
                  dateBased:
                    type: boolean
                  timeBased:
                    type: boolean
            zero:
              type: boolean
        description:
          type: string
        couleur:
          type: string
        periodique:
          type: boolean
        periode:
          type: object
          properties:
            years:
              type: integer
              format: int32
            months:
              type: integer
              format: int32
            days:
              type: integer
              format: int32
            negative:
              type: boolean
            units:
              type: array
              items:
                type: object
                properties:
                  durationEstimated:
                    type: boolean
                  dateBased:
                    type: boolean
                  timeBased:
                    type: boolean
            chronology:
              type: object
              properties:
                id:
                  type: string
                calendarType:
                  type: string
            zero:
              type: boolean
        emploiTemps:
          $ref: '#/components/schemas/EmploiTemps'
    RefreshTokenSalt:
      type: object
      properties:
        id:
          type: integer
          format: int64
        value:
          type: string
        owner:
          $ref: '#/components/schemas/Utilisateur'
    Token:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        expiredAt:
          type: string
          format: date-time
        value:
          type: string
        type:
          type: string
          enum:
          - VERIFY_ACCOUNT
          - RESET_PASSWORD
        owner:
          $ref: '#/components/schemas/Utilisateur'
    Utilisateur:
      type: object
      properties:
        id:
          type: integer
          format: int64
        pseudo:
          type: string
        email:
          type: string
        mdp:
          type: string
          writeOnly: true
        verifie:
          type: boolean
          writeOnly: true
        refreshTokenSalt:
          $ref: '#/components/schemas/RefreshTokenSalt'
        tokens:
          type: array
          writeOnly: true
          items:
            $ref: '#/components/schemas/Token'
        emploisTemps:
          type: array
          writeOnly: true
          items:
            $ref: '#/components/schemas/EmploiTemps'
    EventForm:
      type: object
      properties:
        dateDebut:
          type: string
          format: date-time
          writeOnly: true
        duree:
          type: object
          properties:
            seconds:
              type: integer
              format: int64
            nano:
              type: integer
              format: int32
            negative:
              type: boolean
            units:
              type: array
              items:
                type: object
                properties:
                  durationEstimated:
                    type: boolean
                  dateBased:
                    type: boolean
                  timeBased:
                    type: boolean
            zero:
              type: boolean
          writeOnly: true
        description:
          type: string
          writeOnly: true
        couleur:
          type: string
          writeOnly: true
        periodique:
          type: boolean
          writeOnly: true
        periode:
          type: object
          properties:
            years:
              type: integer
              format: int32
            months:
              type: integer
              format: int32
            days:
              type: integer
              format: int32
            negative:
              type: boolean
            units:
              type: array
              items:
                type: object
                properties:
                  durationEstimated:
                    type: boolean
                  dateBased:
                    type: boolean
                  timeBased:
                    type: boolean
            chronology:
              type: object
              properties:
                id:
                  type: string
                calendarType:
                  type: string
            zero:
              type: boolean
          writeOnly: true
        event:
          $ref: '#/components/schemas/Evenement'
...
    CriteresRerchercheEvents:
      required:
      - dateDebut
      - dateFin
      type: object
      properties:
        dateDebut:
          type: string
          format: date-time
        dateFin:
          type: string
          format: date-time
        description:
          type: string
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
...
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
    PageEvenement:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        pageable:
          $ref: '#/components/schemas/Pageable'
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Evenement'
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean